udmf : disable Hexen polyobj things (require ZDoom ones) in UDMF maps.
There was already ambiguity between the Hexen polyobj things and two very
common Doom monsters (IMP and DEMON).  With UDMF, there is also ambiguity
with line type 1, which is a very common door in Doom/Boom/Heretic but a
explicit polyobj marker in Hexen/ZDoom.

This commit implements a basic solution: in a UDMF map, require the use
of ZDoom polyobj things.